+++ crates/entrypoint/src/cli/command/azure/mod.rs
@@ -2,6 +2,7 @@ pub mod audit;
 pub mod azure_command;
 pub mod group;
 pub mod pim;
+pub mod entra;
 pub mod policy;
 pub mod role;
 pub mod tag;